requirements.txt
flask==2.3.3
flask-cors==4.0.0
torch==2.1.0
torchvision==0.16.0
Pillow==10.0.1
numpy==1.24.3
vecs==0.4.1
psycopg2-binary==2.9.9
python-dotenv==1.0.0
tqdm==4.66.1
git+https://github.com/openai/CLIP.git